Changed floor height conditions for solutions
Previously anything that started above the lava was allowed. Now, to make things more compatible with squish cancel, we restrict to things between 100 and 150 units above the lava. The upper bound could be made a little tighter, but I think this will filter out the majority of bad solutions for now.