Added check to 10k PU route.
It was possible for the position after 1 frame of movement to be on the other platform triangle. This could cause problems if the triangles had different y normals. So added a check to ensure that the y normal is what we expect it to be. This, unfortunately, was responsible for generating some false solutions.