Added checks to confirm that the output stream is open
If you give an invalid file name, or something, then the stream will fail silently. This means that the user could lose solutions without realising it.

Though, this could also be invoked deliberately if the user actually didn't want to output to a file. So, the program will only give a warning if this happens.