Changed internal memory limits to input parameters
Previously set these with #define directives, however this means that the values must be set at compile time and cannot be specified by the user at runtime. This is a holdover from when the brute forcer stored sub-solutions in static arrays. Now that we use dynamic memory allocation, this is no longer necessary.