Optimised search for breakdance solutions
Previously the brute forcer would try every possible stick position. This lead to a lot of duplicates, since a lot of stick positions with maximum magnitude are effectively equivalent. Now only checks stick positions with unique magnitude and yaw. 

This speeds things up by about 3x. Given that this was the slowest part of the brute forcer, this is a pretty significant improvement.